+++ src/StardewModdingAPI/Events/EventArgsMouseStateChanged.cs
@@ -1,5 +1,7 @@
 ﻿using System;
+using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Input;
+using StardewValley;
 
 namespace StardewModdingAPI.Events
 {
@@ -10,10 +12,16 @@ namespace StardewModdingAPI.Events
         ** Accessors
         *********/
         /// <summary>The previous mouse state.</summary>
-        public MouseState NewState { get; private set; }
+        public MouseState PriorState { get; private set; }
 
         /// <summary>The current mouse state.</summary>
-        public MouseState PriorState { get; private set; }
+        public MouseState NewState { get; private set; }
+
+        /// <summary>The previous mouse position on the screen adjusted for the zoom level.</summary>
+        public Point PriorPosition { get; private set; }
+
+        /// <summary>The current mouse position on the screen adjusted for the zoom level.</summary>
+        public Point NewPosition { get; private set; }
 
 
         /*********
@@ -22,10 +30,14 @@ namespace StardewModdingAPI.Events
         /// <summary>Construct an instance.</summary>
         /// <param name="priorState">The previous mouse state.</param>
         /// <param name="newState">The current mouse state.</param>
-        public EventArgsMouseStateChanged(MouseState priorState, MouseState newState)
+        /// <param name="priorPosition">The previous mouse position on the screen adjusted for the zoom level.</param>
+        /// <param name="newPosition">The current mouse position on the screen adjusted for the zoom level.</param>
+        public EventArgsMouseStateChanged(MouseState priorState, MouseState newState, Point priorPosition, Point newPosition)
         {
             this.PriorState = priorState;
             this.NewState = newState;
+            this.PriorPosition = priorPosition;
+            this.NewPosition = newPosition;
         }
     }
 }